chore(internal): consolidate TypeScript SDK test helpers into @fern-typescript/test-utils
Extract duplicated test helper functions from 17 test files into the shared
@fern-typescript/test-utils package:

- createMockReference(name): shared Reference mock (was in 8 files)
- createMockZurgSchema(exprText), createMockZurgObjectSchema(): shared Zurg mocks (was in 5 files)
- primitiveTypeRefNode, optionalTypeRefNode, readWriteTypeRefNode, namedTypeRefNode: shared TypeReferenceNode builders (was in 4 files)
- serializeStatements(statements): shared statement serializer (was in 5 files)

Net: -205 lines of duplicated code across 17 test files.
Only replaced helpers with identical implementations; intentionally kept
variant implementations in TypeReferenceConverters.test.ts,
WebsocketTypeSchemaGenerator.test.ts (different createMockZurgSchema), and
type-generator tests (different serializeStatements signature).
