+++ generators/typescript/model/type-generator/src/__test__/GeneratedObjectTypeImpl.test.ts
@@ -4,7 +4,10 @@ import {
     casingsGenerator,
     createDeclaredTypeName,
     createNameAndWireValue,
-    createObjectProperty
+    createObjectProperty,
+    optionalTypeRefNode,
+    primitiveTypeRefNode,
+    readWriteTypeRefNode
 } from "@fern-typescript/test-utils";
 import { Project, ts } from "ts-morph";
 import { assert, describe, expect, it } from "vitest";
@@ -23,108 +26,6 @@ function createFernFilepath(): FernIr.FernFilepath {
         allParts: [casingsGenerator.generateName("types")],
         packagePath: [casingsGenerator.generateName("types")],
         file: casingsGenerator.generateName("types")
-    };
-}
-
-/**
- * Builds a TypeReferenceNode for a primitive type (non-optional).
- */
-function primitiveTypeRefNode(typeName: string): TypeReferenceNode {
-    const typeNode = ts.factory.createKeywordTypeNode(
-        typeName === "string"
-            ? ts.SyntaxKind.StringKeyword
-            : typeName === "number"
-              ? ts.SyntaxKind.NumberKeyword
-              : typeName === "boolean"
-                ? ts.SyntaxKind.BooleanKeyword
-                : ts.SyntaxKind.AnyKeyword
-    );
-    return {
-        isOptional: false,
-        typeNode,
-        typeNodeWithoutUndefined: typeNode,
-        requestTypeNode: undefined,
-        requestTypeNodeWithoutUndefined: undefined,
-        responseTypeNode: undefined,
-        responseTypeNodeWithoutUndefined: undefined
-    };
-}
-
-/**
- * Builds a TypeReferenceNode for an optional type (T | undefined).
- */
-function optionalTypeRefNode(typeName: string): TypeReferenceNode {
-    const baseNode = ts.factory.createKeywordTypeNode(
-        typeName === "string"
-            ? ts.SyntaxKind.StringKeyword
-            : typeName === "number"
-              ? ts.SyntaxKind.NumberKeyword
-              : typeName === "boolean"
-                ? ts.SyntaxKind.BooleanKeyword
-                : ts.SyntaxKind.AnyKeyword
-    );
-    const unionNode = ts.factory.createUnionTypeNode([
-        baseNode,
-        ts.factory.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword)
-    ]);
-    return {
-        isOptional: true,
-        typeNode: unionNode,
-        typeNodeWithoutUndefined: baseNode,
-        requestTypeNode: undefined,
-        requestTypeNodeWithoutUndefined: undefined,
-        responseTypeNode: undefined,
-        responseTypeNodeWithoutUndefined: undefined
-    };
-}
-
-/**
- * Builds a TypeReferenceNode with separate request/response type variants.
- */
-function readWriteTypeRefNode(opts: {
-    baseType: string;
-    requestType?: string;
-    responseType?: string;
-    isOptional?: boolean;
-}): TypeReferenceNode {
-    const baseNode = ts.factory.createTypeReferenceNode(opts.baseType);
-    const reqNode = opts.requestType ? ts.factory.createTypeReferenceNode(opts.requestType) : undefined;
-    const respNode = opts.responseType ? ts.factory.createTypeReferenceNode(opts.responseType) : undefined;
-
-    if (opts.isOptional) {
-        const unionNode = ts.factory.createUnionTypeNode([
-            baseNode,
-            ts.factory.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword)
-        ]);
-        return {
-            isOptional: true,
-            typeNode: unionNode,
-            typeNodeWithoutUndefined: baseNode,
-            requestTypeNode: reqNode
-                ? ts.factory.createUnionTypeNode([
-                      reqNode,
-                      ts.factory.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword)
-                  ])
-                : undefined,
-            requestTypeNodeWithoutUndefined: reqNode,
-            responseTypeNode: respNode
-                ? ts.factory.createUnionTypeNode([
-                      respNode,
-                      ts.factory.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword)
-                  ])
-                : undefined,
-            responseTypeNodeWithoutUndefined: respNode
-        };
-    }
-
-    return {
-        isOptional: false,
-        typeNode: baseNode,
-        typeNodeWithoutUndefined: baseNode,
-        requestTypeNode: reqNode,
-        requestTypeNodeWithoutUndefined: reqNode,
-        responseTypeNode: respNode,
-        responseTypeNodeWithoutUndefined: respNode
     };
 }
