+++ generators/typescript/sdk/client-class-generator/src/__test__/GeneratedStreamingEndpointImplementation.test.ts
@@ -1,6 +1,6 @@
 import { FernIr } from "@fern-fern/ir-sdk";
 import { getTextOfTsNode, PackageId } from "@fern-typescript/commons";
-import { createHttpEndpoint } from "@fern-typescript/test-utils";
+import { createHttpEndpoint, serializeStatements } from "@fern-typescript/test-utils";
 import { ts } from "ts-morph";
 import { assert, describe, expect, it } from "vitest";
 import type { GeneratedEndpointRequest } from "../endpoint-request/GeneratedEndpointRequest.js";
@@ -10,10 +10,6 @@ import { GeneratedStreamingEndpointImplementation } from "../endpoints/Generated
 // ──────────────────────────────────────────────────────────────────────────────
 // Helpers
 // ──────────────────────────────────────────────────────────────────────────────
-
-function serializeStatements(statements: ts.Statement[]): string {
-    return statements.map((s) => getTextOfTsNode(s)).join("\n");
-}
 
 function serializeNodes(nodes: ts.Node[]): string {
     return nodes.map((n) => getTextOfTsNode(n)).join("\n");
