+++ generators/typescript/sdk/sdk-error-schema-generator/src/__test__/SdkErrorSchemaGenerator.test.ts
@@ -1,6 +1,6 @@
 import { FernIr } from "@fern-fern/ir-sdk";
-import { getTextOfTsNode, Reference, Zurg } from "@fern-typescript/commons";
-import { casingsGenerator } from "@fern-typescript/test-utils";
+import { getTextOfTsNode } from "@fern-typescript/commons";
+import { casingsGenerator, createMockReference, createMockZurgSchema } from "@fern-typescript/test-utils";
 import { Project, ts } from "ts-morph";
 import { describe, expect, it } from "vitest";
 
@@ -10,45 +10,6 @@ import { SdkErrorSchemaGenerator } from "../SdkErrorSchemaGenerator.js";
 // ────────────────────────────────────────────────────────────────────────────
 // Helpers
 // ────────────────────────────────────────────────────────────────────────────
-
-function createMockReference(name: string): Reference {
-    return {
-        getExpression: () => ts.factory.createIdentifier(name),
-        getTypeNode: () => ts.factory.createTypeReferenceNode(name),
-        getEntityName: () => ts.factory.createIdentifier(name)
-        // biome-ignore lint/suspicious/noExplicitAny: test mock
-    } as any;
-}
-
-function createMockZurgSchema(exprText: string): Zurg.Schema {
-    const base: Zurg.BaseSchema = {
-        isOptional: false,
-        isNullable: false,
-        toExpression: () => ts.factory.createIdentifier(exprText)
-    };
-    return {
-        ...base,
-        parse: (raw: ts.Expression) => raw,
-        json: (parsed: ts.Expression) => parsed,
-        parseOrThrow: (raw: ts.Expression) =>
-            ts.factory.createCallExpression(
-                ts.factory.createPropertyAccessExpression(ts.factory.createIdentifier(exprText), "parseOrThrow"),
-                undefined,
-                [raw]
-            ),
-        jsonOrThrow: (parsed: ts.Expression) =>
-            ts.factory.createCallExpression(
-                ts.factory.createPropertyAccessExpression(ts.factory.createIdentifier(exprText), "jsonOrThrow"),
-                undefined,
-                [parsed]
-            ),
-        nullable: () => createMockZurgSchema(`${exprText}.nullable()`),
-        optional: () => createMockZurgSchema(`${exprText}.optional()`),
-        optionalNullable: () => createMockZurgSchema(`${exprText}.optionalNullable()`),
-        transform: () => createMockZurgSchema(`${exprText}.transform()`)
-        // biome-ignore lint/suspicious/noExplicitAny: test mock
-    } as any;
-}
 
 function createMockSdkContext() {
     const project = new Project({ useInMemoryFileSystem: true });
